refactor(agents): split prepared model configured facts
